fix warning

+++ Engine/RHI/Vulkan/public/vulkanBuffer.h
@@ -46,7 +46,7 @@ namespace Homura
         VulkanVertexBuffer(VulkanDevicePtr device, VkDeviceSize size, void* pData)
             : VulkanBuffer(device, size, VK_BUFFER_USAGE_TRANSFER_DST_BIT | VK_BUFFER_USAGE_VERTEX_BUFFER_BIT, VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT)
         {
-            updateBufferByStaging(pData, size);
+            updateBufferByStaging(pData, static_cast<uint32_t>(size));
         }
     };
 
@@ -55,7 +55,7 @@ namespace Homura
         VulkanIndexBuffer(VulkanDevicePtr device, VkDeviceSize size, void* pData)
                 : VulkanBuffer(device, size, VK_BUFFER_USAGE_TRANSFER_DST_BIT | VK_BUFFER_USAGE_INDEX_BUFFER_BIT, VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT)
         {
-            updateBufferByStaging(pData, size);
+            updateBufferByStaging(pData, static_cast<uint32_t>(size));
         }
     };
 
@@ -66,7 +66,7 @@ namespace Homura
         {
             if (pData)
             {
-                updateBufferByStaging(pData, size);
+                updateBufferByStaging(pData, static_cast<uint32_t>(size));
             }
         }
     };
@@ -78,7 +78,7 @@ namespace Homura
         {
             if (pData)
             {
-                updateBufferByStaging(pData, size);
+                updateBufferByStaging(pData, static_cast<uint32_t>(size));
             }
         }
     };
